Add option to let parent downtimes suppress child notifications
Old NAgios supported just setting Downtimes on the host and service
notifications would not get sent anymore.

Add the option `ignore_downtime` to the dependencies to also check
whether the parent is in downtime. Hence we can not only consider state
changes for reachability, but also the downtime status.

This can be used to downtime the host only, but suppress notifications
for both host and services and make it unnecessary to downtimes all
services.

Whenever a host is downtimed we have the expectation to not receive
any notifications for the services.

Defaults to `ignore_downtime = true` to keep the old behaviour.

Fixes https://github.com/Icinga/icinga2/issues/3935