Fix for stats min operator
m_Min was initialzed with a value of 0.
The previous fix with a fixed initialization value was not perfect in all cases, e. g. in some cases when no value is returned it will return 9999999999.
This fix will keep the initial value of 0 and will set it to value in case m_Min is 0.