Change last hard state logic
Changed logic how the last hard state is calculated.
Previously, the LastHardState was set to the current state when a hard state change was encountered, so the LastHardState always showed the current hard state.
Now, LastHardState always shows the last hard state, even when the current state is a hard one.
Fixes #5441