changelog.py: Adjust categories and labels: Enhancement, Bug, ITL, Documentation, Support
This PR allows to adjust the categories and matching labels. If no label
matches, "Support" is the old hardcoded default.

Issues with the "Documentation" and "ITL" label will be put into their
respective category at the bottom before "Support". This increases readability
and allows users to focus on the core vs additional config and docs.

"Support" also applies to labels such as "code-quality", "Tests" and "Packages"/"Installation".

Note: Labels are case sensitive. The order of the categories dictionary is
important too.

Since issues and PRs are the same for GitHub, and we don't require an issue
for a PR anymore, we sometimes have duplicates. This patch adds an inline
label called "PR" to highlight these PRs in the Changelog.