edit Icinga license info so that GitHub recognizes it
GitHub uses a library called Licensee to identify a project's license
type. It shows this information in the status bar and via the API if it
can unambiguously identify the license.

This commit modifies a few of Icinga's license-related docs so that
Licensee is able to recognize the repository's license type. It moves
the contents of COPYING.Exceptions to the relevant section of the
README, which has been renamed to "Licensing". It then removes the
COPYING.Exceptions file and updates CMakeLists.txt accordingly.

Collectively, these changes allow Licensee to successfully identify the
license type of Icinga as GPL 2.0.

Signed-off-by: Andrea Kao <eirinikos@gmail.com>