Fix reload on Windows
Again... How does this work? The Service on Windows will now be installed
with "restart service" as recovery behaviour. That means it will
automatically restart when it dies with a non-standard exit code. Upon
reload we run a config check and if that is sucessful, we die and let the
Service Manager handle the restart.