Raise .NET target version to 4.6
This sets the target .NET framework version to 4.6.

The Prefer32Bit option was introduced with .NET 4. When set to true this
will generate x86 machine code on a 64-bit system, but this only affects
projects where the platform target is set to Any CPU. CMake generates
the project based on the system. Since the Prefer32Bit option was auto
generated during the switch to the new .NET version I ensured this is
explicitly set to false.

refs #7090