Drop .NET 2.0/3.0/3.5 as supported .NET platform
This drops .NET 2.0, 3.0 and 3.5 as as supported .NET runtime platform.

The maximum version can only be v4.0, which targets currently all .NET
4.x releases. Therefor I added the sku flag to enforce .NET 4.6 as
minimum runtime platform. Since .NET 4.x is backwards compatible we
should be safe to only set the minimum version/sku here (i.e. we should
be able to run the application on a system with .NET 4.7)

refs #7090