Windows: Dynamically link Boost DLLs (avoid static linking) and ship them in sbin/
This comes with the motivation to avoid bundling Boost context
into our main binary. Whenever the context library allocates
stack memory, or does something weird, we want to see this
in the stack traces as origin from an external DLL and not the
main icinga2.exe binary.

TL;DR - this doesn't fix #7431 but makes debugging and development
a bit more easy. And it follows the Windows approach with DLLs.