Performing a reload might result in zombie processes
If the reload timeout approaches, checkercomponent will set a flag for this situation and running processes will be killed. checkercomponent will wait up to 10 seconds until there are no pending checks anymore. Check results of processes that got killed during this phase will not be processed to avoid false positives.