Fixed name issue with haproxy
haproxy is using the in build status page and haproxy_status is using the socket. So I'd like to propose the change of both like haproxy now is getting haproxy_status since it is using status page and haproxy_status now to haproxy since it is using the real socket for it.