Set log_duration=0 for Host/Service Command Endpoints
This happens in memory after the configuration has been
loaded, and as such, before any connection attempt or sync.

This needs a special explanation in the docs.

If the value already is 0, no objectlock/set is fired. This
avoids extra round trips for e.g. Director generated agent
endpoint objects.