Introduce a passive_interval attribute for checkables
Introduce a passive_interval for Host/Service objects so a user can
(a) express that passive checks are actually used (unfortunately, enable_passive_checks defaults to true) and
(b) give a hint for the interval passive results are noremally expected to arrive in.

Use that information to
(1) not re-schedule active checks direct parents on recovery if passive checks are actually in use
(2) decide whether a checkable is likely to be re-checked soon before fireing supressed notifications.

Also, simplyfy IsLikelyToBeCheckedSoon and avoid a ten-second-interval after recovery where spurious problem notifications could be sent.

Not sure what to do with the GetNextUpdate() method.

Added new fields to MySQL/PostgreSQL schema files, schema upgrade missing.