docs: better instructions for installing wixtoolset on Windows
If you try to follow the existing instructions, you might be greeted by
a long and useless error message when installing this which boils down
to the fact that the DotNet3.5 Chocolatey package actually requires a
reboot but does not properly ask for it. Enable-WindowsOptionalFeature
basically does the same as the Chocolatey package would do, but nicely
asks for a reboot after which the Chocolatey package can be installed
(which basically is a noop then) properly.