Update 09-object-types.md (#7953)
* Update 09-object-types.md

Because in this document it is not clear which units are allowed for DB Cleanup, I asked in the Community for it: https://community.icinga.com/t/db-ido-cleanup-possible-units/3854/2. We came to the conclusion that it would help if there is a notice which units are possible/allowed for this. For better finding I wrote  also a link to the language reference.

* Update 09-object-types.md

changed to the proposed "language reference"

* Update 09-object-types.md

Done. Sorry overlooked

* Update 09-object-types.md

accidentally inserted return undone . Hope now fits 😃

* Update 09-object-types.md

Requested Changes don

* Update 09-object-types.md

changed link to duration-literals