Alternative to just install monitoring-plugins on openSUSE
As alternative to install just the monitoring-plugins on openSUSE and SLE, users can also install the `monitoring-plugins-all` package, which requires monitoring-plugins and additional packages from the server:monitoring repository. These additional packages are just recommended, so - in case a dependency can not be fulfilled - which allows to get more (useful) plugins installed in the first run.