Fix HTTP status for actions with empty match set
If we requested an API action with a valid JSON filter that matches
for zero objects

{"type": "Service", "filter": "service.name==\"doesnt-exist\""}

the actions API responded with

HTTP/1.1 500 Internal Server Error
{"results":[]}

This is semanticlly incorrect, because the server did exactly and
successfully what the user requested, namely to apply an action to zero
objects. In practice such empty filter match sets are common if you use
wildcards or greater/less comparisons in filters.

Therefore return 200 - Ok with an empty result list in such a case.