Docs: remove duration from schedule-downtime example
The API call does not explicitly specify `fixed` so the default of `true` is used. For a fixed downtime, `duration` is ignored so there is no reason for specifying it and it's just confusing ("why do I have to give start/end *and* a duration which should be the difference between start and end anyways?").