StreamLogger: Get rid of from static mutex & cleanup codes
This fixes the problem described in the referenced issue, as it now just hold a lock on itself instead
of static mutex when logging to a console. It's the case that when starting a process, e.g. to validate
Icinga2 configs, and calling `exit(3)` after validation succeeds, then the static `ProcessEntry()` method
from `StreamLogger` class is used to process the entries. See `Logger::~Logger()` for details. However, the
problem here is that after the process has entered to exit(3), work queue threads still log something, and
of course a `static mutex` from the stream logger class is used to ensure `std::cout` thread-safety. But this
static mutex, in this case, may already have been destroyed. And since we don't have a handler for this use
case (to lock a dead mutex), it will directly call `std::terminate()`, as it immediately raises a system error.
Our unhandled exception handler of course tries to generate statcktrace from this situation but, for some reason
(could be due to the same reasone as above) the static variables in `exception.cpp` such as `l_lastExceptionObj`
etc.. aren't set, which in turn causes a segfault in `RethrowUncaughtException()`.